+++ src/app/app-routing.module.ts
@@ -1,10 +1,32 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { ProductComponent } from './feature/product/components/product/product.component';
+import { AboutComponent, PathNotFoundComponent } from './layout';
 
-const routes: Routes = [];
+
+const routes: Routes = [
+  {
+    path: 'about',
+    component: AboutComponent
+  },
+  {
+    path: '',
+    redirectTo: '/home',
+    pathMatch: 'full'
+  },
+  {
+    // The router will match this route if the URL requested
+    // doesn't match any paths for routes defined in our configuration
+    path: '**',
+    component: PathNotFoundComponent
+  }
+];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [
+    RouterModule.forRoot(routes)
+  ],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
+