+++ src/app/feature/product/components/product-list/product-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, Output, EventEmitter} from '@angular/core';
 
 import { ProductModel } from '../../models/product.model';
 import { ProductService } from '../../../shared/services/products.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-product-list',
@@ -13,7 +14,8 @@ export class ProductListComponent implements OnInit {
   @Output() buyProd: EventEmitter<any> = new EventEmitter();
 
   products: Promise<Array<ProductModel>>;
-  constructor(private productsService: ProductService) { }
+  constructor(private productsService: ProductService,
+    private router: Router) { }
 
   ngOnInit() {
     this.products = this.productsService.getProducts();
@@ -21,5 +23,9 @@ export class ProductListComponent implements OnInit {
 
   onBuy(prod: ProductModel): void {
     this.buyProd.emit(prod);
+  }
+  onDetails(prod: ProductModel): void {
+    const link = ['/details', prod];
+    this.router.navigate(link);
   }
 }