+++ src/app/feature/product/product.module.ts
@@ -2,11 +2,11 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ProductService } from '../shared/services/products.service';
-import { ProductListComponent } from './components/product-list/product-list.component';
-import { ProductComponent } from './components/product/product.component';
+import { ProductComponent,ProductListComponent } from './components';
+import { ProductsRoutingModule } from './products-routing.module';
 
 @NgModule({
-  imports: [CommonModule],
+  imports: [CommonModule,ProductsRoutingModule],
   declarations: [ProductComponent, ProductListComponent],
   providers: [ProductService],
   exports: [ProductComponent, ProductListComponent]