+++ src/app/feature/product/products-routing.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { ProductListComponent, ProductComponent } from './components';
+
+const routes: Routes = [
+  {
+    path: 'home',
+    component: ProductListComponent
+  },{
+    path:'details/:product',
+    component: ProductComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ProductsRoutingModule { }