+++ src/app/layout/components/index.ts
@@ -1,0 +1,2 @@
+export * from './about/about.component';
+export * from './path-not-found/path-not-found.component';